#16 Abstrakce XML IO handleru. Ten jediný potřeboval být více abstraktní, než doposud byl.
