+++ app/src/main/java/cz/zcu/fav/remotestimulatorcontrol/io/GenericXMLHandler.java
@@ -1,0 +1,117 @@
+package cz.zcu.fav.remotestimulatorcontrol.io;
+
+import android.util.Xml;
+
+import org.xmlpull.v1.XmlPullParserException;
+import org.xmlpull.v1.XmlPullParserFactory;
+import org.xmlpull.v1.XmlSerializer;
+
+import java.io.IOException;
+
+/**
+ * Abstraktní třída pro uložení a načtení dat ve formátu XML
+ */
+public abstract class GenericXMLHandler implements IOHandler {
+
+    // region Constants
+    // Výchozí očekávané kódování XML souboru
+    public static final String DEFAULT_ENCODING = "UTF-8";
+    public static final boolean STAND_ALONE = true;
+    // Namespace
+    public static final String NAMESPACE = "";
+    // endregion
+
+    // region Variables
+    private static boolean _minify = false;
+    // endregion
+
+    // region Static methods
+    /**
+     * Nastaví minifikaci XML dokumentu
+     *
+     * @param minify True, pokud se má xml dokument minifikovat, jinak false
+     */
+    public static void setMinify(boolean minify) {
+        _minify = minify;
+    }
+    // endregion
+
+    // region Public methods
+    /**
+     * Vytvoří nový XML serializer s automatickým zalamováním řádků
+     *
+     * @return {@link XmlSerializer}
+     */
+    protected XmlSerializer createSerializer() {
+        XmlSerializer serializer = Xml.newSerializer();
+        serializer.setFeature("http://xmlpull.org/v1/doc/features.html#indent-output", !_minify);
+
+        return serializer;
+    }
+
+    /**
+     * Vytvoří novou XmlPullParserFactory s podporou namespacu
+     *
+     * @return {@link XmlPullParserFactory}
+     * @throws XmlPullParserException
+     */
+    protected XmlPullParserFactory createFactory() throws XmlPullParserException {
+        XmlPullParserFactory factory = XmlPullParserFactory.newInstance();
+        factory.setNamespaceAware(true);
+
+        return factory;
+    }
+
+    /**
+     * Vytvoří a zapíše nový XML tag
+     *
+     * @param serializer {@link XmlSerializer} Serializer, do kterého se zapisují hodnoty
+     * @param tag Tag
+     * @param value Hodnota
+     * @throws IOException Pokud něco nevyjde
+     */
+    protected void writeTag(XmlSerializer serializer, String tag, int value) throws IOException {
+        writeTag(serializer, tag, String.valueOf(value));
+    }
+
+    /**
+     * Vytvoří a zapíše nový XML tag
+     *
+     * @param serializer {@link XmlSerializer} Serializer, do kterého se zapisují hodnoty
+     * @param tag Tag
+     * @param value Hodnota
+     * @throws IOException Pokud něco nevyjde
+     */
+    protected void writeTag(XmlSerializer serializer, String tag, double value) throws IOException {
+        writeTag(serializer, tag, String.valueOf(value));
+    }
+
+    /**
+     * Vytvoří a zapíše nový XML tag
+     *
+     * @param serializer {@link XmlSerializer} Serializer, do kterého se zapisují hodnoty
+     * @param tag Tag
+     * @param value Hodnota
+     * @throws IOException Pokud něco nevyjde
+     */
+    protected void writeTag(XmlSerializer serializer, String tag, String value) throws IOException {
+        writeTag(serializer, NAMESPACE, tag, value);
+    }
+
+    /**
+     * Vytvoří a zapíše nový XML tag
+     *
+     * @param serializer {@link XmlSerializer} Serializer, do kterého se zapisují hodnoty
+     * @param namespace Namespace
+     * @param tag Tag
+     * @param value Hodnota
+     * @throws IOException Pokud něco nevyjde
+     */
+    protected void writeTag(XmlSerializer serializer, String namespace, String tag, String value) throws IOException {
+        serializer.startTag(namespace, tag);
+        serializer.text(value);
+        serializer.endTag(namespace, tag);
+    }
+    // endregion
+
+}
